requirements.txt
requests>=2.0.0
PyYAML>=6.0
pytest>=8.0.0
pytest-mock>=3.0.0